+++ lib/Connection.js
@@ -20,8 +20,8 @@ const {
   iXmlNodeHead, iXmlNodeScriptOpen, iXmlNodeScriptClose,
 } = require('./ixml');
 
-const { iRestHttp } = require('./irest');
-const { db2Call } = require('./istoredp');
+const { iRestHttp } = require('./transports/irest');
+const { db2Call } = require('./transports/istoredp');
 
 const availableTransports = {
   idb: db2Call,
